+++ app/pages/changelog.vue
@@ -14,6 +14,50 @@ useSeoMeta({
 
 const versions = [
   {
+    title: "1.3.0",
+    date: "2026-03-15",
+    content: `
+### Features
+
+- Add a contact form to the website for users to submit feedback and report bugs
+- Add a changelog page and an faq page to the game to answer user questions and keep them informed
+
+### Changes
+
+- Changed the colors used in app to be more consistent
+- Updated the look of different buttons and components in the app to be closer to Warframe's design style
+- Added signature weapon and unique rolling animation as new grid game categories
+- Added icons to links to indicate if they go to an external site
+
+### Bug Fixes
+
+- Fixed images not having defined width and height attributes
+- Fixed some cards disappearing from legacy game stats when the game mode is unlimited
+- Fixed some colors changing when using colorblind mode that should be consistent across all modes
+`,
+    badge: "v1.3.0",
+  },
+  {
+    title: "1.2.3",
+    date: "2026-02-21",
+    content: `
+### Fixes
+
+- Fixed share stats button not working in classic and ability game modes
+`,
+    badge: "v1.2.3",
+  },
+  {
+    title: "1.2.3",
+    date: "2026-02-21",
+    content: `
+### Fixes
+
+- Fixed share stats button not working in classic and ability game modes
+`,
+    badge: "v1.2.3",
+  },
+  {
     title: "1.2.2",
     date: "2026-02-20",
     content: `